Moved tests to modules
