+++ NeuroNotes.slnx
@@ -7,28 +7,26 @@
     <Project Path="src/AiAssistant/NeuroNotes.AiAssistant.Infrastructure/NeuroNotes.AiAssistant.Infrastructure.csproj" />
     <Project Path="src/AiAssistant/NeuroNotes.AiAssistant.Persistence/NeuroNotes.AiAssistant.Persistence.csproj" />
     <Project Path="src/AiAssistant/NeuroNotes.AiAssistant.Public/NeuroNotes.AiAssistant.Public.csproj" />
+    <Project Path="src/AiAssistant/NeuroNotes.AiAssistant.UnitTests/NeuroNotes.AiAssistant.UnitTests.csproj" />
   </Folder>
   <Folder Name="/src/AudioProcessing/">
     <Project Path="src/AudioProcessing/NeuroNotes.AudioProcessing.Application/NeuroNotes.AudioProcessing.Application.csproj" />
     <Project Path="src/AudioProcessing/NeuroNotes.AudioProcessing.Infrastructure/NeuroNotes.AudioProcessing.Infrastructure.csproj" />
     <Project Path="src/AudioProcessing/NeuroNotes.AudioProcessing.Public/NeuroNotes.AudioProcessing.Public.csproj" />
+    <Project Path="src/AudioProcessing/NeuroNotes.AudioProcessing.UnitTests/NeuroNotes.AudioProcessing.UnitTests.csproj" />
   </Folder>
   <Folder Name="/src/GitHub/">
     <Project Path="src/GitHub/NeuroNotes.GitHub.Application/NeuroNotes.GitHub.Application.csproj" />
     <Project Path="src/GitHub/NeuroNotes.GitHub.Infrastructure/NeuroNotes.GitHub.Infrastructure.csproj" />
     <Project Path="src/GitHub/NeuroNotes.GitHub.Persistence/NeuroNotes.GitHub.Persistence.csproj" />
     <Project Path="src/GitHub/NeuroNotes.GitHub.Public/NeuroNotes.GitHub.Public.csproj" />
+    <Project Path="src/GitHub/NeuroNotes.GitHub.UnitTests/NeuroNotes.GitHub.UnitTests.csproj" />
   </Folder>
   <Folder Name="/src/TelegramBot/">
     <Project Path="src/TelegramBot/NeuroNotes.TelegramBot.Application/NeuroNotes.TelegramBot.Application.csproj" />
     <Project Path="src/TelegramBot/NeuroNotes.TelegramBot.Infrastructure/NeuroNotes.TelegramBot.Infrastructure.csproj" />
     <Project Path="src/TelegramBot/NeuroNotes.TelegramBot.Persistence/NeuroNotes.TelegramBot.Persistence.csproj" />
     <Project Path="src/TelegramBot/NeuroNotes.TelegramBot.Public/NeuroNotes.TelegramBot.Public.csproj" />
-  </Folder>
-  <Folder Name="/tests/">
-    <Project Path="tests/NeuroNotes.AiAssistant.UnitTests/NeuroNotes.AiAssistant.UnitTests.csproj" />
-    <Project Path="tests/NeuroNotes.AudioProcessing.UnitTests/NeuroNotes.AudioProcessing.UnitTests.csproj" />
-    <Project Path="tests/NeuroNotes.GitHub.UnitTests/NeuroNotes.GitHub.UnitTests.csproj" />
-    <Project Path="tests/NeuroNotes.TelegramBot.UnitTests/NeuroNotes.TelegramBot.UnitTests.csproj" />
+    <Project Path="src/TelegramBot/NeuroNotes.TelegramBot.UnitTests/NeuroNotes.TelegramBot.UnitTests.csproj" />
   </Folder>
 </Solution>
