+++ src/AiAssistant/NeuroNotes.AiAssistant.Persistence/GlobalUsings.cs
@@ -4,7 +4,7 @@ global using Microsoft.EntityFrameworkCore;
 global using Microsoft.EntityFrameworkCore.Design;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Options;
-global using NeuroNotes.AiAssistant.Persistence.Configurations;
+global using NeuroNotes.AiAssistant.Persistence.DbContexts;
 global using NeuroNotes.AiAssistant.Persistence.Entities;
 global using NeuroNotes.AiAssistant.Persistence.Repositories;
 global using NeuroNotes.AiAssistant.Public.Interfaces;
