+++ src/AiAssistant/NeuroNotes.AiAssistant.UnitTests/UnitTest1.cs
