+++ src/TelegramBot/NeuroNotes.TelegramBot.UnitTests/NeuroNotes.TelegramBot.UnitTests.csproj
@@ -4,7 +4,7 @@
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
         <OutputType>Exe</OutputType>
-        <RootNamespace>NeuroNotes.GitHub.UnitTests</RootNamespace>
+        <RootNamespace>NeuroNotes.TelegramBot.UnitTests</RootNamespace>
         <TargetFramework>net10.0</TargetFramework>
     </PropertyGroup>
 
@@ -23,9 +23,9 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\src\GitHub\NeuroNotes.GitHub.Application\NeuroNotes.GitHub.Application.csproj" />
-        <ProjectReference Include="..\..\src\GitHub\NeuroNotes.GitHub.Persistence\NeuroNotes.GitHub.Persistence.csproj" />
-        <ProjectReference Include="..\..\src\GitHub\NeuroNotes.GitHub.Public\NeuroNotes.GitHub.Public.csproj" />
+      <ProjectReference Include="..\NeuroNotes.TelegramBot.Application\NeuroNotes.TelegramBot.Application.csproj" />
+      <ProjectReference Include="..\NeuroNotes.TelegramBot.Infrastructure\NeuroNotes.TelegramBot.Infrastructure.csproj" />
+      <ProjectReference Include="..\NeuroNotes.TelegramBot.Persistence\NeuroNotes.TelegramBot.Persistence.csproj" />
     </ItemGroup>
 
 </Project>
